deploy/frontend/requirements.txt
streamlit==1.45.0
requests==2.32.3
Pillow==10.2.0
